requirements.txt
absl-py==1.4.0
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==2.0.12
dataclasses==0.8
google-auth==2.20.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.4
importlib-metadata==4.8.3
Markdown==3.3.7
numpy==1.19.5
oauthlib==3.2.2
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.1.1
urllib3==1.26.16
Werkzeug==2.0.3
zipp==3.6.0
